+++ benchmark/matmul/benchmark_matmul_intrinsic.py
@@ -11,7 +11,7 @@ from tilelang.intrinsics import get_swizzle_layout
 from tilelang.intrinsics.mma_macro_generator import (
     TensorCoreIntrinEmitter,)
 from tilelang.transform import simplify_prim_func
-from tilelang.autotuner import autotune, jit
+from tilelang.autotuner import autotune
 import itertools
 
 # Configure logger
@@ -267,12 +267,8 @@ def matmul(M,
         warmup=3,
         rep=5,
     )
-    @jit(
+    @tl.jit(
         out_idx=[2],
-        supply_type=tl.TensorSupplyType.Integer,
-        ref_prog=ref_program,
-        skip_check=True,
-        target="auto",
     )
     def kernel(
         block_row_warps=None,
