+++ benchmark/matmul_fp8/benchmark_matmul.py
@@ -5,9 +5,9 @@ import argparse
 import itertools
 import logging
 
-import tilelang as tl
 import tilelang.language as T
-from tilelang.autotuner import autotune, jit
+from tilelang.autotuner import autotune
+from tilelang import jit
 
 # Configure logger
 logger = logging.getLogger(__name__)
@@ -165,10 +165,6 @@ def matmul(M, N, K, with_roller):
     )
     @jit(
         out_idx=[2],
-        supply_type=tl.TensorSupplyType.Integer,
-        ref_prog=ref_program,
-        skip_check=True,
-        target="auto",
     )
     def kernel(
         block_M=None,
