+++ examples/flash_attention/example_mha_fwd_bhsd.py
@@ -9,7 +9,7 @@ import tilelang.language as T
 import itertools
 import argparse
 from functools import partial
-
+from tilelang import jit
 
 def get_configs():
     block_M = [128]
@@ -156,7 +156,7 @@ def flashattn(batch, heads, seq_q, seq_kv, dim, is_causal, tune=False):
     if tune:
 
         @autotune(configs=get_configs(), warmup=10, rep=10)
-        @jit(out_idx=[3], supply_type=tilelang.TensorSupplyType.Integer, ref_prog=None)
+        @jit(out_idx=[3])
         def kernel(block_M=None, block_N=None, num_stages=None, threads=None):
             return kernel_func(block_M, block_N, num_stages, threads)
