+++ examples/linear_attention/example_mamba_chunk_scan.py
@@ -199,7 +199,7 @@ def chunk_scan_fwd(batch, seqlen, chunk_size, ngroups, nheads, headdim, dstate, 
     if tune:
 
         @autotune(configs=get_configs(), warmup=10, rep=10)
-        @jit(out_idx=[7], supply_type=tilelang.TensorSupplyType.Normal, ref_prog=None)
+        @tilelang.jit(out_idx=[7])
         def kernel(block_M=None,
                    block_N=None,
                    block_K=None,
